+++ src/datasource/DronerDatasource.ts
@@ -48,11 +48,12 @@ export class DronerDatasource {
   }
 
   static updateDroner(data: DronerEntity): Promise<any> {
-    for (let i = 0; data.dronerDrone.length > i; i++) {
-      let checkIdDronerDrone = data.dronerDrone[i].id;
-      checkIdDronerDrone == "" && delete data.dronerDrone[i].dronerId;
-      delete data.dronerDrone[i].id;
-    }
+    // for (let i = 0; data.dronerDrone.length > i; i++) {
+    //   let checkIdDronerDrone = data.dronerDrone[i].id;
+    //   checkIdDronerDrone == "" && delete data.dronerDrone[i].dronerId;
+    //   delete data.dronerDrone[i].id;
+    // }
+    delete data["dronerDrone"];
 
     if (data.dronerArea.provinceId == 0) {
       delete data.dronerArea["provinceId"];
@@ -61,7 +62,6 @@ export class DronerDatasource {
       delete data.dronerArea["subdistrictId"];
       delete data.dronerArea["distance"];
     }
-    console.log("base", data);
     return httpClient
       .patch(BASE_URL + "/droner/" + data.id, data)
       .then((response) => {