+++ src/datasource/UploadImageDatasource.ts
@@ -1,7 +1,6 @@
 import { BASE_URL, httpClient } from "../config/develop-config";
 import { ImageEntity } from "../entities/UploadImageEntities";
 
-
 export class UploadImageDatasouce {
   static uploadImage(file: ImageEntity): Promise<any> {
     const formDataProfile = new FormData();
@@ -37,6 +36,20 @@ export class UploadImageDatasouce {
     };
     return httpClient
       .delete(BASE_URL + "/file/delete", { params })
+      .then((response) => {
+        return response.data;
+      })
+      .catch((err) => {
+        console.log(err, "err getImg");
+      });
+  }
+
+  static getImageFinish(path: string): Promise<any> {
+    const params = {
+      path: path,
+    };
+    return httpClient
+      .get(BASE_URL + "/tasks/file/geturl", { params })
       .then((response) => {
         return response.data;
       })