+++ src/pages/droner/DetailWorkDroner.tsx
@@ -1,63 +1,80 @@
-import Layout from "../../../../components/layout/Layout";
+import Layout from "../../components/layout/Layout";
 import React, { useEffect, useState } from "react";
-import {
-  Avatar,
-  Badge,
-  Empty,
-  Form,
-  Input,
-  Row,
-  Select,
-  Tag,
-  Upload,
-} from "antd";
-import { BackIconButton } from "../../../../components/button/BackButton";
-import { CardContainer } from "../../../../components/card/CardContainer";
-import { CardHeader } from "../../../../components/header/CardHearder";
-import color from "../../../../resource/color";
-import GoogleMap from "../../../../components/map/GoogleMap";
-import { LAT_LNG_BANGKOK } from "../../../../definitions/Location";
+import { Avatar, Badge, Form, Input, Row, Select, Tag } from "antd";
+import { BackIconButton } from "../../components/button/BackButton";
+import { CardContainer } from "../../components/card/CardContainer";
+import { CardHeader } from "../../components/header/CardHearder";
+import color from "../../resource/color";
+import GoogleMap from "../../components/map/GoogleMap";
+import { LAT_LNG_BANGKOK } from "../../definitions/Location";
 import TextArea from "antd/lib/input/TextArea";
+import moment from "moment";
+import {
+  taskDetailEntity,
+  taskDetailEntity_INIT,
+} from "../../entities/DronerRankEntities";
+import { DronerRankDatasource } from "../../datasource/DronerRankDatasource";
 import {
   AddressEntity,
   AddressEntity_INIT,
-} from "../../../../entities/AddressEntities";
+  FullAddressEntiry_INIT,
+  FullAddressEntity,
+} from "../../entities/AddressEntities";
 import {
   StarFilled,
   CalendarOutlined,
   ClockCircleOutlined,
 } from "@ant-design/icons";
+import { UploadImageDatasouce } from "../../datasource/UploadImageDatasource";
 import {
-  DetailFinishTask,
-  DetailFinishTask_INIT,
-} from "../../../../entities/TaskFinishEntities";
-import { TaskFinishedDatasource } from "../../../../datasource/TaskFinishDatasource";
-import moment from "moment";
-import { UploadImageDatasouce } from "../../../../datasource/UploadImageDatasource";
+  DistrictEntity,
+  DistrictEntity_INIT,
+  ProviceEntity,
+  ProvinceEntity_INIT,
+  SubdistrictEntity,
+  SubdistrictEntity_INIT,
+} from "../../entities/LocationEntities";
+import { LocationDatasource } from "../../datasource/LocationDatasource";
+import { CropPurposeSprayEntity } from "../../entities/CropEntities";
 const _ = require("lodash");
 let queryString = _.split(window.location.search, "=");
 const dateFormat = "DD/MM/YYYY";
 const timeFormat = "HH:mm";
 
-function FinishTasks() {
+function DetailWorkDroner() {
   const taskId = queryString[1];
-  const [data, setData] = useState<DetailFinishTask>(DetailFinishTask_INIT);
+  const [data, setData] = useState<taskDetailEntity>(taskDetailEntity_INIT);
+  let imgList: (string | boolean)[] = [];
   const [mapPosition, setMapPosition] = useState<{ lat: number; lng: number }>({
     lat: LAT_LNG_BANGKOK.lat,
     lng: LAT_LNG_BANGKOK.lng,
   });
-  const fetchDetailTask = async () => {
-    await TaskFinishedDatasource.getDetailFinishTaskById(taskId).then((res) => {
+  const [imgFinish, setImgFinish] = useState<any>();
+
+  const fetchTask = async () => {
+    await DronerRankDatasource.getTaskDetail(taskId).then((res) => {
       setData(res);
+      console.log(res)
       setMapPosition({
-        lat: parseFloat(res.data.farmerPlot.lat),
-        lng: parseFloat(res.data.farmerPlot.long),
+        lat: parseFloat(res.farmerPlot.lat),
+        lng: parseFloat(res.farmerPlot.long),
       });
+      let getImgFinish = res.imagePathFinishTask;
+      imgList.push(getImgFinish != null ? getImgFinish : "");
+      var i = 0;
+      for (i; imgList.length > i; i++) {
+        i == 0 &&
+          UploadImageDatasouce.getImageFinish(imgList[i].toString()).then(
+            (resImg) => {
+              setImgFinish(resImg.url);
+            }
+          );
+      }
     });
   };
 
   useEffect(() => {
-    fetchDetailTask();
+    fetchTask();
   }, []);
 
   const formatCurrency = (e: any) => {
@@ -67,10 +84,12 @@ function FinishTasks() {
     });
   };
   const onPreviewImg = async () => {
-    let src = data.imageTaskUrl;
+    let src = imgFinish;
     if (!src) {
       src = await new Promise((resolve) => {
         const reader = new FileReader();
+        reader.readAsDataURL(imgFinish);
+        reader.onload = () => resolve(reader.result);
       });
     }
     const image = new Image();
@@ -78,7 +97,6 @@ function FinishTasks() {
     const imgWindow = window.open(src);
     imgWindow?.document.write(image.outerHTML);
   };
-
   const starIcon = (
     <StarFilled
       style={{
@@ -88,7 +106,6 @@ function FinishTasks() {
       }}
     />
   );
-
   const renderAppointment = (
     <Form style={{ padding: "32px" }}>
       <div className="row">
@@ -99,7 +116,7 @@ function FinishTasks() {
               <Form.Item>
                 <Input
                   style={{ width: "100%" }}
-                  value={moment(new Date(data.data.dateAppointment)).format(
+                  value={moment(new Date(data.dateAppointment)).format(
                     dateFormat
                   )}
                   disabled
@@ -112,7 +129,7 @@ function FinishTasks() {
               <Form.Item>
                 <Input
                   style={{ width: "100%" }}
-                  value={moment(new Date(data.data.dateAppointment)).format(
+                  value={moment(new Date(data.dateAppointment)).format(
                     timeFormat
                   )}
                   disabled
@@ -126,8 +143,8 @@ function FinishTasks() {
             <Select
               disabled
               value={
-                data.data.purposeSpray !== null
-                  ? data.data.purposeSpray.purposeSprayName
+                data.purposeSpray !== null
+                  ? data.purposeSpray.purposeSprayName
                   : "-"
               }
             />
@@ -135,13 +152,13 @@ function FinishTasks() {
           <label>เป้าหมายการฉีดพ่น</label>
           <Form.Item>
             <span style={{ color: color.Grey }}>
-              {data.data.targetSpray !== null ? data.data.targetSpray : "-"}
+              {data.targetSpray !== null ? data.targetSpray : "-"}
             </span>
           </Form.Item>
           <label>การเตรียมยา</label>
           <Form.Item>
             <span style={{ color: color.Grey }}>
-              {data.data.preparationBy !== null ? data.data.preparationBy : "-"}
+              {data.preparationBy !== null ? data.preparationBy : "-"}
             </span>
           </Form.Item>
           <label>ภาพงานจากนักบินโดรน</label>
@@ -150,15 +167,12 @@ function FinishTasks() {
             <div
               className="hiddenFileInput"
               style={{
-                backgroundImage: `url(${data.imageTaskUrl})`,
-                display:
-                  data.imageTaskUrl !== null
-                    ? `url(${data.imageTaskUrl})`
-                    : undefined,
+                backgroundImage: `url(${imgFinish})`,
+                display: imgFinish != null ? `url(${imgFinish})` : undefined,
               }}
             ></div>
             <div className="ps-5">
-              {data.imageTaskUrl !== null ? (
+              {imgFinish != undefined && (
                 <>
                   <Tag
                     color={color.Success}
@@ -168,7 +182,7 @@ function FinishTasks() {
                     View
                   </Tag>
                 </>
-              ) : null}
+              )}
             </div>
           </div>
 
@@ -176,7 +190,7 @@ function FinishTasks() {
           <label>หมายเหตุ</label>
           <Form.Item>
             <span style={{ color: color.Grey }}>
-              {data.data.statusRemark !== null ? data.data.statusRemark : "-"}
+              {data.statusRemark ? data.statusRemark : "-"}
             </span>
           </Form.Item>
         </div>
@@ -185,47 +199,44 @@ function FinishTasks() {
           <label>ค่าบริการ</label>
           <Form.Item style={{ color: color.Grey }}>
             <span>
-              {data.data.price !== null
-                ? formatCurrency(data.data.price) + " " + "บาท"
+              {" "}
+              {data.price !== null
+                ? formatCurrency(data.price) + " " + "บาท"
                 : "0.00" + " " + "บาท"}
             </span>{" "}
             <span>
-              {"(จำนวน" + " " + data.data.farmAreaAmount + " " + "ไร่)"}
+              {data.farmAreaAmount !== null
+                ? "(จำนวน" + " " + data.farmAreaAmount + " " + "ไร่)"
+                : "0"}
             </span>
           </Form.Item>
           <br />
           <Form.Item>
             <div className="row">
-              <div className="col-lg-3">คะแนนรีวิว </div>
+              <label className="col-lg-3">คะแนนรีวิว </label>
               <div className="col-lg-6">
-                {" "}
-                <span>
-                  {data.data.reviewDronerAvg > "0" ? (
-                    <Row>
-                      {starIcon}
-                      <span>
-                        {parseFloat(data.data.reviewDronerAvg).toFixed(1)}
-                      </span>
-                    </Row>
-                  ) : (
-                    "-"
-                  )}
-                </span>
+                {data.reviewDronerAvg > "0" ? (
+                  <Row>
+                    {starIcon}
+                    <span>{parseFloat(data.reviewDronerAvg).toFixed(1)}</span>
+                  </Row>
+                ) : (
+                  "-"
+                )}
               </div>
             </div>
-          </Form.Item>
-          <Form.Item>
+            <br />
             <div className="row">
               <div className="col-lg-6" style={{ color: color.Grey }}>
                 1. มารยาทนักบิน{" "}
               </div>
               <div className="col-lg-6">
-                {data.data.reviewDronerDetail !== null ? (
+                {data.reviewDronerDetail !== null ? (
                   <Row>
                     {starIcon}
                     <span>
                       {parseFloat(
-                        data.data.reviewDronerDetail.pilotEtiquette
+                        data.reviewDronerDetail.pilotEtiquette
                       ).toFixed(1)}
                     </span>
                   </Row>
@@ -239,13 +250,13 @@ function FinishTasks() {
                 2. ความตรงต่อเวลา{" "}
               </div>
               <div className="col-lg-6">
-                {data.data.reviewDronerDetail !== null ? (
+                {data.reviewDronerDetail !== null ? (
                   <Row>
                     {starIcon}
                     <span>
-                      {parseFloat(
-                        data.data.reviewDronerDetail.punctuality
-                      ).toFixed(1)}
+                      {parseFloat(data.reviewDronerDetail.punctuality).toFixed(
+                        1
+                      )}
                     </span>
                   </Row>
                 ) : (
@@ -258,12 +269,12 @@ function FinishTasks() {
                 3. ความเชี่ยวชาญในการพ่น{" "}
               </div>
               <div className="col-lg-6">
-                {data.data.reviewDronerDetail !== null ? (
+                {data.reviewDronerDetail !== null ? (
                   <Row>
                     {starIcon}
                     <span>
                       {parseFloat(
-                        data.data.reviewDronerDetail.sprayExpertise
+                        data.reviewDronerDetail.sprayExpertise
                       ).toFixed(1)}
                     </span>
                   </Row>
@@ -278,8 +289,8 @@ function FinishTasks() {
               rows={3}
               disabled
               value={
-                data.data.reviewDronerDetail != null
-                  ? data.data.reviewDronerDetail.comment
+                data.reviewDronerDetail != null
+                  ? data.reviewDronerDetail.comment
                   : undefined
               }
             />
@@ -288,7 +299,7 @@ function FinishTasks() {
           <Form.Item>
             <span style={{ color: color.Success }}>
               <Badge color={color.Success} />
-              {data.data.status == "DONE" ? "เสร็จสิ้น" : null}
+              {data.status == "DONE" ? "เสร็จสิ้น" : null}
               <br />
             </span>
           </Form.Item>
@@ -297,8 +308,8 @@ function FinishTasks() {
     </Form>
   );
   const renderFarmer = (
-    <Form key={data.data.farmerId}>
-      <div className="container text-center" style={{ padding: "30px" }}>
+    <Form key={data.farmerId}>
+      <div style={{ padding: "30px" }}>
         <div className="row">
           <div className="col-lg-4 text-start">
             <label>ชื่อ-นามสกุล</label>
@@ -306,7 +317,7 @@ function FinishTasks() {
               <Input
                 disabled
                 defaultValue={
-                  data.data.farmer.firstname + " " + data.data.farmer.lastname
+                  data.farmer.firstname + " " + data.farmer.lastname
                 }
               />
             </Form.Item>
@@ -314,7 +325,7 @@ function FinishTasks() {
           <div className="col-lg-4 text-start">
             <label>เบอร์โทร</label>
             <Form.Item>
-              <Input disabled defaultValue={data.data.farmer.telephoneNo} />
+              <Input disabled defaultValue={data.farmer.telephoneNo} />
             </Form.Item>
           </div>
         </div>
@@ -322,13 +333,13 @@ function FinishTasks() {
           <div className="col-lg-4 text-start">
             <label>แปลง</label>
             <Form.Item>
-              <Select disabled defaultValue={data.data.farmerPlot.plotName} />
+              <Select disabled defaultValue={data.farmerPlot.plotName} />
             </Form.Item>
           </div>
           <div className="col-lg-4 text-start">
             <label>พืชที่ปลูก</label>
             <Form.Item>
-              <Input disabled defaultValue={data.data.farmerPlot.plantName} />
+              <Input disabled defaultValue={data.farmerPlot.plantName} />
             </Form.Item>
           </div>
           <div className="col-lg-4 text-start">
@@ -336,7 +347,7 @@ function FinishTasks() {
             <Form.Item>
               <Input
                 disabled
-                defaultValue={data.data.farmerPlot.raiAmount}
+                defaultValue={data.farmerPlot.raiAmount}
                 suffix="ไร่"
               />
             </Form.Item>
@@ -349,12 +360,12 @@ function FinishTasks() {
               <Input
                 disabled
                 defaultValue={
-                  data.data.farmerPlot.plotArea !== null
-                    ? data.data.farmerPlot.plotArea.subdistrictName +
+                  data.farmerPlot.plotArea !== null
+                    ? data.farmerPlot.plotArea.subdistrictName +
                       "/" +
-                      data.data.farmerPlot.plotArea.districtName +
+                      data.farmerPlot.plotArea.districtName +
                       "/" +
-                      data.data.farmerPlot.plotArea.provinceName
+                      data.farmerPlot.plotArea.provinceName
                     : "-"
                 }
               />
@@ -386,7 +397,7 @@ function FinishTasks() {
           <div className="col-lg-12 text-start">
             <label>จุดสังเกต</label>
             <Form.Item>
-              <Input disabled value={data.data.farmerPlot.landmark} />
+              <Input disabled value={data.farmerPlot.landmark} />
             </Form.Item>
           </div>
         </div>
@@ -398,52 +409,46 @@ function FinishTasks() {
       <div className="row">
         <div className="col-lg-3">
           <span>
-            {data.data.droner.firstname + " " + data.data.droner.lastname}
+            {data.droner.firstname + " " + data.droner.lastname}
             <br />
             <p style={{ fontSize: "12px", color: color.Grey }}>
-              {data.data.droner.dronerCode}
+              {data.droner.dronerCode}
             </p>
           </span>
         </div>
         <div className="col-lg-3">
-          <span>{data.data.droner.telephoneNo}</span>
+          <span>{data.droner.telephoneNo}</span>
         </div>
         <div className="col-lg-4">
-          <span>
-            {" "}
-            {data.data.droner.address !== null
-              ? data.data.droner.address.subdistrict.subdistrictName +
-                "," +
-                " " +
-                data.data.droner.address.district.districtName +
-                "," +
-                " " +
-                data.data.droner.address.province.region
-              : "-"}
-          </span>
+          {data.droner.address.subdistrict.subdistrictName +
+            "," +
+            " " +
+            data.droner.address.district.districtName +
+            "," +
+            " " +
+            data.droner.address.province.region}
         </div>
         <div className="col-lg">
           <span>
             <Avatar
               size={25}
               src={
-                data.data.droner.dronerDrone[0] != null
-                  ? data.data.droner.dronerDrone[0].drone.droneBrand
-                      .logoImagePath
+                data.droner.dronerDrone && data.droner.dronerDrone[0] != null
+                  ? data.droner.dronerDrone[0].drone.droneBrand.logoImagePath
                   : null
               }
               style={{ marginRight: "5px" }}
             />
-            {data.data.droner.dronerDrone[0] != null
-              ? data.data.droner.dronerDrone[0].drone.droneBrand.name
+            {data.droner.dronerDrone && data.droner.dronerDrone[0] != null
+              ? data.droner.dronerDrone[0].drone.droneBrand.name
               : "-"}
           </span>
           <br />
-          <p style={{ fontSize: "12px", color: color.Grey }}>
-            {data.data.droner.dronerDrone.length > 1
+          <span style={{ color: color.Grey, fontSize: "12px" }}>
+            {data.droner.dronerDrone && data.droner.dronerDrone.length! > 1
               ? "(มากกว่า 1 ยี่ห้อ)"
               : null}
-          </p>
+          </span>
         </div>
       </div>
     </Form>
@@ -458,8 +463,8 @@ function FinishTasks() {
                 ยอดรวมค่าบริการ (หลังรวมค่าธรรมเนียม)
                 <br />
                 <b style={{ fontSize: "20px", color: color.Success }}>
-                  {data.data.totalPrice !== null
-                    ? formatCurrency(data.data.totalPrice) + " " + "บาท"
+                  {data.totalPrice !== null
+                    ? formatCurrency(data.totalPrice) + " " + "บาท"
                     : "0.00" + " " + "บาท"}
                 </b>
               </span>
@@ -469,13 +474,11 @@ function FinishTasks() {
         <div className="row">
           <div className="col-lg-4">
             <Form.Item>
-              <label>ค่าบริการ (ก่อนคิดค่าคำนวณ)</label>
+              <label>ค่าบริการ (ก่อนคิดค่าธรรมเนียม)</label>
               <Input
                 disabled
                 value={
-                  data.data.price !== null
-                    ? formatCurrency(data.data.price) + " " + "บาท"
-                    : "0.00" + " " + "บาท"
+                  data.price !== null ? formatCurrency(data.price) : "0.00"
                 }
                 suffix="บาท"
               />
@@ -483,14 +486,14 @@ function FinishTasks() {
           </div>
           <div className="col-lg-4">
             <Form.Item>
-              <label>ค่าธรรมเนียม (คิด 5% ของราคารวม)</label>
+              <label>ค่าธรรมเนียม (5% ของค่าบริการ)</label>
               <Input
                 disabled
                 placeholder="0.0"
                 value={
-                  data.data.discountFee !== null
-                    ? parseFloat(data.data.discountFee).toFixed(2) + " " + "บาท"
-                    : "0.00" + " " + "บาท"
+                  data.discountFee !== null
+                    ? formatCurrency(data.discountFee)
+                    : "0.00"
                 }
                 suffix="บาท"
               />
@@ -501,11 +504,7 @@ function FinishTasks() {
               <label>ส่วนลดค่าธรรมเนียม</label>
               <Input
                 disabled
-                value={
-                  data.data.fee !== null
-                    ? formatCurrency(data.data.fee) + " " + "บาท"
-                    : "0.00" + " " + "บาท"
-                }
+                value={data.fee !== null ? formatCurrency(data.fee) : "0.00"}
                 suffix="บาท"
               />
             </Form.Item>
@@ -519,11 +518,13 @@ function FinishTasks() {
     <Layout>
       <Row>
         <BackIconButton
-          onClick={() => (window.location.href = "/IndexFinishTask")}
+          onClick={() =>
+            (window.location.href = "/DetailRankDroner?=" + data.dronerId)
+          }
         />
         <span className="pt-4">
           <strong style={{ fontSize: "20px" }}>
-            รายละเอียดงาน #{data.data.taskNo}
+            รายละเอียดงาน #{data.taskNo}
           </strong>
         </span>
       </Row>
@@ -550,4 +551,4 @@ function FinishTasks() {
   );
 }
 
-export default FinishTasks;
+export default DetailWorkDroner;